Cross-platform SSL.Trusted
