--- modules/tests/js/src/test/scala/SslTestPlatform.scala
@@ -1,7 +1,0 @@
-// Copyright (c) 2018-2021 by Rob Norris
-// This software is licensed under the MIT License (MIT).
-// For more information see LICENSE or https://opensource.org/licenses/MIT
-
-package tests
-
-trait SslTestPlatform {}
