+++ modules/tests/shared/src/test/scala/SslTest.scala
@@ -5,18 +5,26 @@
 package tests
 
 import cats.effect._
-import cats.syntax.all._
 import fs2.io.net.tls.SSLException
 import natchez.Trace.Implicits.noop
 import skunk._
-import fs2.CompositeFailure
 
-class SslTest extends ffstest.FTest with SslTestPlatform {
+class SslTest extends ffstest.FTest {
 
   object Port {
     val Invalid = 5431
     val MD5     = 5432
     val Trust   = 5433
+  }
+
+  test("successful login with SSL.Trusted (ssl available)") {
+    Session.single[IO](
+      host     = "localhost",
+      user     = "jimmy",
+      database = "world",
+      password = Some("banana"),
+      ssl      = SSL.Trusted,
+    ).use(_ => IO.unit)
   }
 
   test("successful login with SSL.None (ssl available)") {
@@ -36,12 +44,27 @@ class SslTest extends ffstest.FTest with SslTestPlatform {
       database = "world",
       password = Some("banana"),
       ssl      = SSL.System,
+    ).use(_ => IO.unit).assertFailsWith[SSLException].as("sigh") // TODO! Better failure!
+  }
+
+  test("failed login with SSL.Trusted (ssl not available)") {
+    Session.single[IO](
+      host     = "localhost",
+      user     = "postgres",
+      database = "world",
+      ssl      = SSL.Trusted,
+      port     = Port.Trust
+    ).use(_ => IO.unit).assertFailsWith[Exception].as("ok") // TODO! Better failure!
+  }
+
+  test("successful login with SSL.Trusted.withFallback(true) (ssl not available)") {
+    Session.single[IO](
+      host     = "localhost",
+      user     = "postgres",
+      database = "world",
+      ssl      = SSL.Trusted.withFallback(true),
+      port     = Port.Trust
     ).use(_ => IO.unit)
-      .adaptError {
-        case ex @ CompositeFailure(head, tail) =>
-          tail.prepend(head).collectFirst({ case ex: SSLException => ex }).getOrElse(ex)
-        }
-      .assertFailsWith[SSLException].as("sigh") // TODO! Better failure!
   }
 
   test("successful login with SSL.None (ssl not available)") {
