+++ modules/tests/shared/src/test/scala/issue/238.scala
